<feat>: change the core event driven logic:
1) Replacing epoll by poll(): even if poll() is slower, it is able to
notify event on filesystems FDs, which is mandatory for us (Ressource
class).

2) Modify the logic behind response: before this commit the response
variable consist only on one std::string, it wasn't the best solution
because it was very slow to pop the front part of the string when it was
consume while sending to the client.
The solution I propose is to use std::queue, a FIFO data structure which,
in one hand, let the Ressources fill the response chunk by chunk (with the
help of the Connexion::pushResponse() function), and in the other hand
let the Connexion send chunk by chunk to the client with the front()
method.

This commit is supposed to fix #14 issue.